QCSXCAD.cpp: isEmpty() instead of null-comparison
Comparing a QString with null was already deprecated in Qt 4.0,
and has been removed in Qt 6. Instead, we use isEmpty() to
check whether a file name is empty. Note that it's not strictly
equivalent with the original null-comparison - a zero-length
"\0"  string is empty but not null.

But logically, we simply want to see if the file name is empty
at high level, regardless of whether the underlying string is
null. So we use isEmpty() instead of isNull().

Signed-off-by: Yifeng Li <tomli@tomli.me>